backend/pyproject.toml
[project]
name = "qa-base-backend"
version = "0.1.0"
description = "Backend for AI-based QA automation tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "google-genai>=1.0.0",
    "browser-use",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv.sources]
browser-use = { path = "../browser_use", editable = true }
